# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Young Tots Edventures Summer Safari 2026 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Registration operations
  - name: reviews
    description: Parent review operations
  - name: admin
    description: Admin operations
  - name: uploads
    description: File upload operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /registrations/count:
    get:
      operationId: getRegistrationCount
      tags: [registrations]
      summary: Get total registration count (public)
      responses:
        "200":
          description: Registration count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationCount"

  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Submit a new registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /registrations/{id}/payment-proof:
    post:
      operationId: uploadPaymentProof
      tags: [uploads]
      summary: Upload M-Pesa payment screenshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Payment proof uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Registration not found

  /reviews:
    get:
      operationId: listPublicReviews
      tags: [reviews]
      summary: List approved reviews (public)
      responses:
        "200":
          description: Approved reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a new review (public, pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted for approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicReview"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reviews:
    get:
      operationId: listReviews
      tags: [admin]
      summary: List all reviews (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
        "401":
          description: Unauthorized

  /admin/reviews/{id}:
    patch:
      operationId: updateReview
      tags: [admin]
      summary: Update a review's content or status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
    delete:
      operationId: deleteReview
      tags: [admin]
      summary: Delete a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials

  /admin/registrations:
    get:
      operationId: listRegistrations
      tags: [admin]
      summary: List all registrations (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: paymentStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [id, parentName, parentEmail, parentPhone, childName, childAge, paymentStatus, createdAt]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationListResponse"
        "401":
          description: Unauthorized

  /admin/registrations/{id}:
    get:
      operationId: getRegistration
      tags: [admin]
      summary: Get a single registration (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found
    delete:
      operationId: deleteRegistration
      tags: [admin]
      summary: Delete a single registration (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found

  /admin/registrations/{id}/payment-status:
    patch:
      operationId: updatePaymentStatus
      tags: [admin]
      summary: Update payment status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentStatusUpdate"
      responses:
        "200":
          description: Updated registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusUpdateResult"
        "404":
          description: Not found

  /admin/registrations/export:
    get:
      operationId: exportRegistrations
      tags: [admin]
      summary: Export registrations as CSV
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: paymentStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string

  /admin/stats:
    get:
      operationId: getStats
      tags: [admin]
      summary: Get registration stats summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResult:
      type: object
      properties:
        id:
          type: integer
      required:
        - id

    RegistrationInput:
      type: object
      required:
        - parentName
        - parentPhone
        - parentEmail
        - childName
        - childDateOfBirth
        - childAge
        - homeAddress
        - emergencyContactName
        - emergencyContactRelationship
        - emergencyContactPhone
        - authorizedPickupPerson
        - authorizedPickupPhone
        - consentAccepted
        - consentSignature
        - consentSignedBy
      properties:
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
        childName:
          type: string
        childDateOfBirth:
          type: string
        childAge:
          type: integer
        homeAddress:
          type: string
        allergies:
          type: ["string", "null"]
        medicalConditions:
          type: ["string", "null"]
        physicalLimitations:
          type: ["string", "null"]
        specialNotes:
          type: ["string", "null"]
        emergencyContactName:
          type: string
        emergencyContactRelationship:
          type: string
        emergencyContactPhone:
          type: string
        emergency2ContactName:
          type: ["string", "null"]
        emergency2ContactRelationship:
          type: ["string", "null"]
        emergency2ContactPhone:
          type: ["string", "null"]
        authorizedPickupPerson:
          type: string
        authorizedPickupPhone:
          type: string
        consentAccepted:
          type: boolean
        consentSignature:
          type: string
        consentSignedBy:
          type: string

    Registration:
      type: object
      required:
        - id
        - parentName
        - parentPhone
        - parentEmail
        - childName
        - childDateOfBirth
        - childAge
        - homeAddress
        - emergencyContactName
        - emergencyContactRelationship
        - emergencyContactPhone
        - authorizedPickupPerson
        - authorizedPickupPhone
        - consentAccepted
        - consentSignature
        - consentSignedBy
        - paymentStatus
        - createdAt
      properties:
        id:
          type: integer
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
        childName:
          type: string
        childDateOfBirth:
          type: string
        childAge:
          type: integer
        homeAddress:
          type: string
        allergies:
          type: ["string", "null"]
        medicalConditions:
          type: ["string", "null"]
        physicalLimitations:
          type: ["string", "null"]
        specialNotes:
          type: ["string", "null"]
        emergencyContactName:
          type: string
        emergencyContactRelationship:
          type: string
        emergencyContactPhone:
          type: string
        emergency2ContactName:
          type: ["string", "null"]
        emergency2ContactRelationship:
          type: ["string", "null"]
        emergency2ContactPhone:
          type: ["string", "null"]
        authorizedPickupPerson:
          type: string
        authorizedPickupPhone:
          type: string
        consentAccepted:
          type: boolean
        consentSignature:
          type: string
        consentSignedBy:
          type: string
        consentTimestamp:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, confirmed, rejected]
        paymentProofUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    RegistrationListResponse:
      type: object
      required:
        - registrations
        - total
        - page
        - limit
      properties:
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/Registration"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    PaymentProofInput:
      type: object
      required:
        - paymentProofBase64
        - mimeType
      properties:
        paymentProofBase64:
          type: string
        mimeType:
          type: string

    PaymentStatusUpdate:
      type: object
      required:
        - paymentStatus
      properties:
        paymentStatus:
          type: string
          enum: [pending, confirmed, rejected]

    PaymentStatusUpdateResult:
      allOf:
        - $ref: "#/components/schemas/Registration"
        - type: object
          properties:
            emailSent:
              type: boolean
            emailError:
              type: ["string", "null"]

    AdminLoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    AuthToken:
      type: object
      required:
        - token
        - username
      properties:
        token:
          type: string
        username:
          type: string

    RegistrationCount:
      type: object
      required:
        - total
        - capacity
        - spotsLeft
      properties:
        total:
          type: integer
        capacity:
          type: integer
        spotsLeft:
          type: integer

    StatsResponse:
      type: object
      required:
        - total
        - pending
        - confirmed
        - rejected
        - todayRegistrations
      properties:
        total:
          type: integer
        pending:
          type: integer
        confirmed:
          type: integer
        rejected:
          type: integer
        todayRegistrations:
          type: integer

    ReviewInput:
      type: object
      required:
        - name
        - email
        - rating
        - review
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 200
        rating:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          minLength: 10
          maxLength: 1000
        website:
          type: string
          description: Honeypot field — must be empty; bots that fill it are rejected silently.

    PublicReview:
      type: object
      required:
        - id
        - name
        - rating
        - review
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        rating:
          type: integer
        review:
          type: string
        createdAt:
          type: string

    PublicReviewListResponse:
      type: object
      required:
        - reviews
        - total
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/PublicReview"
        total:
          type: integer

    Review:
      type: object
      required:
        - id
        - name
        - email
        - rating
        - review
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        rating:
          type: integer
        review:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string

    ReviewListResponse:
      type: object
      required:
        - reviews
        - total
        - page
        - limit
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    ReviewUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        rating:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          minLength: 10
          maxLength: 1000
        status:
          type: string
          enum: [pending, approved, rejected]