# lib/api-zod/src/generated/types/publicReview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Young Tots Edventures Summer Safari 2026 API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicReview {
  id: number;
  name: string;
  rating: number;
  review: string;
  createdAt: string;
}