# lib/api-zod/src/generated/types/publicReviewListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Young Tots Edventures Summer Safari 2026 API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicReview } from './publicReview';

export interface PublicReviewListResponse {
  reviews: PublicReview[];
  total: number;
}