# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Young Tots Edventures Summer Safari 2026 API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewStatus } from './reviewStatus';

export interface Review {
  id: number;
  name: string;
  email: string;
  rating: number;
  review: string;
  status: ReviewStatus;
  createdAt: string;
}