# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Young Tots Edventures Summer Safari 2026 API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 200 */
  email: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  review: string;
  /** Honeypot field — must be empty; bots that fill it are rejected silently. */
  website?: string;
}