# lib/api-zod/src/generated/types/reviewListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Young Tots Edventures Summer Safari 2026 API
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from './review';

export interface ReviewListResponse {
  reviews: Review[];
  total: number;
  page: number;
  limit: number;
}